// src/api/models/orderResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { OrderItemResponse } from './orderItemResponse';

export interface OrderResponse {
  id?: number;
  orderNo?: string;
  userId?: number;
  totalAmount?: number;
  payAmount?: number;
  freightAmount?: number;
  status?: number;
  statusText?: string;
  paymentTime?: string;
  deliveryTime?: string;
  receiveTime?: string;
  comment?: string;
  receiverName?: string;
  receiverPhone?: string;
  receiverAddress?: string;
  createdAt?: string;
  updatedAt?: string;
  items?: OrderItemResponse[];
  totalItems?: number;
  canBeCancelled?: boolean;
  canBeShipped?: boolean;
  canBeCompleted?: boolean;
}